Linux SPM work
Former-commit-id: b66edae99d46e72596a2ce0b37a9cafdf0888899

+++ Sources/cURL/include/p_curl.h
@@ -16,7 +16,7 @@ extern inline CURLcode curl_easy_setopt_cstr(CURL *handle, CURLoption option, co
 	return curl_easy_setopt(handle, option, value);
 }
 
-extern inline CURLcode curl_easy_setopt_int64(CURL *handle, CURLoption option, int64_t value)
+extern inline CURLcode curl_easy_setopt_int64(CURL *handle, CURLoption option, long long value)
 {
 	return curl_easy_setopt(handle, option, value);
 }
